packages/backend/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "alacard-backend"
version = "0.1.0"
description = "FastAPI backend for Alacard notebook generator"
authors = ["Alacard Team"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
celery = {extras = ["redis"], version = "^5.3.4"}
redis = "^5.0.1"
websockets = "^12.0"
httpx = "^0.24.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
ruff = "^0.1.6"

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.black]
line-length = 88